+++ examples/bilinear_reconstruct/facedb/facedb.h
@@ -57,7 +57,7 @@ public:
     static int                                  NumDimIden()    { return gTensorShape[1];     }
     static int                                  NumDimExpr()    { return gTensorShape[2];     }
     static int                                  NumVertices()   { return gTensorShape[0] / 3; }
-    static int                                  NumTriangles()  { return gTriangles.size();   }
+    static int                                  NumTriangles()  { return (int)gTriangles.size();   }
     static const Tensor3 &                      CoreTensor()    { return gCoreTensor;         }
 
     static const MatrixRM &                     IdenUT()        { return gIdenUT; }
